Catch up commit
o Opened threads are remembered across invocations. This uses
  GSettings, so you have to "make install" for it to install the
  schema in the right spot. If the thread 404ed though, you'll get a
  blank ThreadView that you'll have to close manually.

o Implemented a catalog pane.
    You have to enable each board you want to monitor from the
    application menu. But then you can just double-click on a thread
    and it will open.

oo There is an entry bar above it you can type in to search the
   catalog. To get to the next match after you've typed, use the arrow
   keys.

o Added a notify toggle for each thread. This notifies on every post.

o By default, notifications are delivered for posts on threads that
  have been idle for 5 minutes. To disable this, you have to use run:
  # gsettings set com.talisein.fourchan.native.gtk notify-idle false

o Added an autoscroll toggle for each thread.

o Code tags on /g/ are opened in an inline gVim process. If you add or
  remove a row, however, it will break. I think I need to use
  Gtk::Widget::reparent() to fix it.

o Thread (and catalog) scraping now is run in a thread that never
  exits.