Canceller: implementation
When std::function callbacks are passed to objects that are not tied
tot he destruction of the originating object, they may be illegally
involked after destruction.

Canceller should be stored as a std::shared_ptr in every originating
class, and cancel() should be called in the destructor. Whenever you
pass a callback, also pass a canceller. The callback should only be
called via Canceller::involke_if_not_cancelled(callback);